+++ docker-compose.yaml
@@ -2,18 +2,15 @@ version: '3.0'
 
 services:
   grafana:
-    container_name: 'grafana-x-ray-datasource'
     platform: 'linux/amd64'
     build:
-      context: ./.config
       args:
-        grafana_image: ${GRAFANA_IMAGE:-grafana-enterprise}
         grafana_version: ${GRAFANA_VERSION:-latest}
     environment:
       - AWS_ACCESS_KEY_ID=${AWS_ACCESS_KEY_ID}
       - AWS_SECRET_ACCESS_KEY=${AWS_SECRET_ACCESS_KEY}
     ports:
       - 3000:3000/tcp
-    volumes:
-      - ./dist:/var/lib/grafana/plugins/grafana-x-ray-datasource
-      - ./provisioning:/etc/grafana/provisioning
+    extends:
+      file: .config/docker-compose-base.yaml
+      service: grafana
