+++ package.json
@@ -7,7 +7,7 @@
     "dev": "webpack -w -c ./.config/webpack/webpack.config.ts --env development",
     "e2e": "playwright test",
     "e2e:report": "playwright show-report",
-    "lint": "eslint --cache .",
+    "lint": "eslint --cache --ignore-path ./.gitignore --ext .js,.jsx,.ts,.tsx .",
     "lint:fix": "yarn run lint --fix && prettier --write --list-different .",
     "server": "docker compose up --build",
     "sign": "npx --yes @grafana/sign-plugin@latest",
@@ -22,11 +22,12 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@emotion/css": "11.13.5",
-    "@grafana/data": "^12.1.0",
+    "@grafana/data": "^12.2.0",
+    "@grafana/i18n": "^12.2.0",
     "@grafana/plugin-ui": "^0.13.0",
-    "@grafana/runtime": "^12.1.0",
-    "@grafana/schema": "^12.1.0",
-    "@grafana/ui": "^12.1.0",
+    "@grafana/runtime": "^12.2.0",
+    "@grafana/schema": "^12.2.0",
+    "@grafana/ui": "^12.2.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-router-dom": "^6.22.0",
@@ -56,6 +57,8 @@
     "@types/lodash": "4.17.24",
     "@types/node": "24.12.3",
     "@types/prismjs": "1.26.6",
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0",
     "@types/semver": "7.7.1",
     "@types/slate": "0.47.16",
     "@typescript-eslint/eslint-plugin": "8.59.2",
@@ -65,7 +68,6 @@
     "css-loader": "7.1.4",
     "eslint": "10.3.0",
     "eslint-config-prettier": "10.1.8",
-    "eslint-plugin-deprecation": "3.0.0",
     "eslint-plugin-jsdoc": "62.9.0",
     "eslint-plugin-prettier": "5.5.5",
     "eslint-plugin-react": "7.37.5",
@@ -74,27 +76,27 @@
     "fork-ts-checker-webpack-plugin": "9.1.0",
     "glob": "13.0.6",
     "identity-obj-proxy": "3.0.0",
-    "imports-loader": "5.0.0",
+    "imports-loader": "^5.0.0",
     "jest": "30.4.1",
     "jest-environment-jsdom": "30.4.1",
     "lefthook": "2.1.6",
     "prettier": "3.8.3",
     "prismjs": "1.30.0",
-    "replace-in-file-webpack-plugin": "1.0.6",
+    "replace-in-file-webpack-plugin": "^1.0.6",
     "sass": "1.99.0",
     "sass-loader": "16.0.8",
     "semver": "7.7.4",
     "style-loader": "4.0.0",
     "swc-loader": "0.2.7",
     "terser-webpack-plugin": "5.5.0",
-    "ts-node": "10.9.2",
+    "ts-node": "^10.9.2",
     "tsconfig-paths": "4.2.0",
     "typescript": "5.9.3",
     "webpack": "5.106.2",
     "webpack-cli": "7.0.2",
-    "webpack-livereload-plugin": "3.0.2",
-    "webpack-subresource-integrity": "5.1.0",
-    "webpack-virtual-modules": "0.6.2"
+    "webpack-livereload-plugin": "^3.0.2",
+    "webpack-subresource-integrity": "^5.1.0",
+    "webpack-virtual-modules": "^0.6.2"
   },
   "packageManager": "yarn@4.14.1"
 }
